+++ src/components/icons/AnalyticsIcon.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+export const AnalyticsIcon: React.FC<React.SVGProps<SVGSVGElement>> = (
+  props
+) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 -960 960 960"
+    fill="currentColor"
+    aria-hidden
+    focusable="false"
+    {...props}
+  >
+    <path d="M640-160v-280h160v280H640Zm-240 0v-640h160v640H400Zm-240 0v-440h160v440H160Z" />
+  </svg>
+);